Trial_deeptech-saketh/requirements.txt
streamlit
google-generativeai
pymupdf
trafilatura
sentence-transformers
torch
plotly
docx2txt